requirements.txt
async-generator==1.10
asyncio==3.4.3
attrs==21.2.0
certifi==2021.10.8
cffi==1.15.0
chardet==4.0.0
click==8.0.3
cryptography==35.0.0
Flask==2.0.2
fyers-apiv2==2.0.5
h11==0.12.0
idna==2.10
itsdangerous==2.0.1
Jinja2==3.0.2
MarkupSafe==2.0.1
outcome==1.1.0
pycparser==2.21
pyOpenSSL==21.0.0
redis==3.5.3
requests==2.25.1
selenium==4.0.0
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
tornado==6.1
trio==0.19.0
trio-websocket==0.9.2
urllib3==1.26.7
websocket-client==1.2.1
websockets==8.1
Werkzeug==2.0.2
wsproto==1.0.0
